+++ popkorn/src/commonTest/kotlin/ClientTests.kt
@@ -23,11 +23,11 @@ internal class ClientTests : PopKornTest() {
 
         val classApp = injector.inject(TestClassByApp::class, "app")
         val classUse = injector.inject(TestClassByUse::class, "use")
-        val classNew = injector.inject(TestClassByNew::class,"new")
-        val classManual = injector.inject(TestClassNoProvider::class,"some")
+        val classNew = injector.inject(TestClassByNew::class, "new")
+        val classManual = injector.inject(TestClassNoProvider::class, "some")
 
-        val interfaceApp = injector.inject(TestInterface::class,"app")
-        val interfaceUse = injector.inject(TestInterface::class,"use")
+        val interfaceApp = injector.inject(TestInterface::class, "app")
+        val interfaceUse = injector.inject(TestInterface::class, "use")
         val interfaceNew = injector.injectNullable(TestInterface::class, "new")
 
         assertSame(classApp, interfaceApp)