+++ popkorn/src/iosMain/kotlin/InjectorObjC.kt
@@ -60,14 +60,14 @@ class InjectorObjC(private val injector: InjectorController) {
     // If want to specify the environment, do it with Environment("value")
     fun create(clazz: ObjCClass, vararg assistedInstances: Any): Any {
         val environment = assistedInstances.singleOrNull { it is Environment } as? Environment
-        val parameters = assistedInstances.filterNot { it is Environment }.map { if (it is InstanceObjC) it.toKotlin() else it }
-        return injector.create(clazz.kotlinClass(), parameters, environment?.value)
+        val factory = assistedInstances.filterNot { it is Environment }.toFactory()
+        return injector.create(clazz.kotlinClass(), environment?.value, factory)
     }
 
     fun create(clazz: ObjCProtocol, vararg assistedInstances: Any): Any {
         val environment = assistedInstances.singleOrNull { it is Environment } as? Environment
-        val parameters = assistedInstances.filterNot { it is Environment }.map { if (it is InstanceObjC) it.toKotlin() else it }
-        return injector.create(clazz.kotlinClass(), parameters, environment?.value)
+        val factory = assistedInstances.filterNot { it is Environment }.toFactory()
+        return injector.create(clazz.kotlinClass(), environment?.value, factory)
     }
 
 
@@ -75,6 +75,12 @@ class InjectorObjC(private val injector: InjectorController) {
 
     fun purge() = injector.purge()
 
+
+    private fun List<Any>.toFactory(): ParametersFactory {
+        return ParametersFactory.Builder().apply {
+            forEach { add(if (it is InstanceObjC) it.toKotlin() else Instance(it)) }
+        }.build()
+    }
 
     private fun InstanceObjC.toKotlin(): Instance<*> {
         return when (this) {