+++ popkorn/src/jsMain/kotlin/InjectorJS.kt
@@ -42,14 +42,20 @@ class InjectorJS(private val injector: InjectorController) {
     // If want to specify the environment, do it with Environment("value")
     fun <T : Any> create(clazz: JsClass<T>, vararg assistedInstances: Any): T {
         val environment = assistedInstances.singleOrNull { it is Environment } as? Environment
-        val parameters = assistedInstances.filterNot { it is Environment }.map { if (it is InstanceJS<*>) it.toKotlin() else it }
-        return injector.create(clazz.kotlinClass(), parameters, environment?.value)
+        val factory = assistedInstances.filterNot { it is Environment }.toFactory()
+        return injector.create(clazz.kotlinClass(), environment?.value, factory)
     }
 
     fun reset() = injector.reset()
 
     fun purge() = injector.purge()
 
+
+    private fun List<Any>.toFactory(): ParametersFactory {
+        return ParametersFactory.Builder().apply {
+            forEach { add(if (it is InstanceJS<*>) it.toKotlin() else Instance(it)) }
+        }.build()
+    }
 
     private fun <T : Any> InstanceJS<T>.toKotlin() = Instance(instance, type.kotlinClass(), environment)