+++ popkorn/src/jvmMain/kotlin/InjectorJVM.kt
@@ -41,8 +41,8 @@ class InjectorJVM(private val injector: InjectorController) {
     // If want to specify the environment, do it with Environment("value")
     fun <T : Any> create(clazz: Class<T>, vararg assistedInstances: Any): T {
         val environment = assistedInstances.singleOrNull { it is Environment } as? Environment
-        val parameters = assistedInstances.filterNot { it is Environment }.map { if (it is InstanceJVM<*>) it.toKotlin() else it }
-        return injector.create(clazz.kotlinClass(), parameters, environment?.value)
+        val factory = assistedInstances.filterNot { it is Environment }.toFactory()
+        return injector.create(clazz.kotlinClass(), environment?.value, factory)
     }
 
 
@@ -50,6 +50,12 @@ class InjectorJVM(private val injector: InjectorController) {
 
     fun purge() = injector.purge()
 
+
+    private fun List<Any>.toFactory(): ParametersFactory {
+        return ParametersFactory.Builder().apply {
+            forEach { add(if (it is InstanceJVM<*>) it.toKotlin() else Instance(it)) }
+        }.build()
+    }
 
     private fun <T : Any> InstanceJVM<T>.toKotlin() = Instance(instance, type.kotlinClass(), environment)